fix(comment): 댓글 개수 타입을 long으로 변경
- CommentRepository에서 특정 게시글의 댓글 수를 반환하는 메서드의 리턴 타입을 int 에서 long 으로 변경

+++ src/main/java/sssdev/tcc/domain/comment/repository/CommentRepository.java
@@ -5,5 +5,5 @@ import sssdev.tcc.domain.comment.domain.Comment;
 
 public interface CommentRepository extends CommentReadRepository, JpaRepository<Comment, Long> {
 
-    int countByPostId(Long postId);
+    long countByPostId(Long postId);
 }
